requirements.txt
streamlit
google-genai
pydub
python-dotenv
yt-dlp
